Refactor; split out find boundaries step.
